Fixed LearningStudent and Supervisor entities and implemented controller and tests

+++ src/main/java/com/kesizo/cetpe/backend/restapi/controller/LearningStudentController.java
@@ -5,7 +5,11 @@ import com.kesizo.cetpe.backend.restapi.service.LearningStudentService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.rest.webmvc.ResourceNotFoundException;
+import org.springframework.http.HttpStatus;
+import org.springframework.transaction.TransactionSystemException;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Map;
@@ -23,36 +27,78 @@ public class LearningStudentController {
         return _learningStudentService.getAllLearningStudents();
     }
 
-    @RequestMapping(value = "/api/cetpe/lstudent/username/{username}", method = RequestMethod.GET)
+    @RequestMapping(value = "/api/cetpe/lstudent/{username}", method = RequestMethod.GET)
     public LearningStudent learningStudentByUsername(@PathVariable String username){
 
-        return _learningStudentService.getLearningStudentByUserName(username);
+        final LearningStudent currentStudent;
+
+        currentStudent = _learningStudentService.getLearningStudentByUserName(username);
+
+        if (currentStudent==null) {
+            throw new ResourceNotFoundException("Student with username= " + username + " not found");
+        }
+        return currentStudent;
     }
 
 
     @PostMapping("/api/cetpe/lstudent")
-    public LearningStudent create(@RequestBody Map<String, String> body) {
+    @ResponseStatus(HttpStatus.CREATED)
+    public LearningStudent create(@RequestBody Map<String, Object> body) {
 
-        String username = body.get("username");
-        String firstName = body.get("firstName");
-        String lastName = body.get("lastName");
+        try {
+            String username = body.get("username").toString();
+            String firstName = body.get("firstName").toString();
+            String lastName = body.get("lastName").toString();
 
-        return _learningStudentService.createLearningStudent(username,firstName, lastName);
+            return _learningStudentService.createLearningStudent(username,firstName, lastName);
+        }
+        catch (NullPointerException np) {
+            logger.warn("Error mandatory parameter null provided when creating student");
+            logger.warn(np.getMessage());
+            throw new ResponseStatusException(
+                    HttpStatus.BAD_REQUEST, "Error mandatory parameter null provided when creating student", np);
+        }
+        catch (TransactionSystemException cve) {
+            logger.warn(cve.getMessage());
+            throw new ResponseStatusException(
+                    HttpStatus.BAD_REQUEST, "Error Bean Validation failed", cve);
+        }
+
+
     }
 
-    @PutMapping("/api/cetpe/lstudent/{id}")
-    public LearningStudent update(@PathVariable String username, @RequestBody Map<String, String> body) {
+    @PutMapping("/api/cetpe/lstudent/{username}")
+    public LearningStudent update(@PathVariable String username, @RequestBody Map<String, Object> body) {
 
-        String firstName = body.get("firstName");
-        String lastName = body.get("lastName");
+        try {
+            String firstName = body.get("firstName").toString();
+            String lastName = body.get("lastName").toString();
 
-        return _learningStudentService.updateLearningStudent(username,firstName, lastName);
+            return _learningStudentService.updateLearningStudent(username,firstName, lastName);
+        }
+        catch (NullPointerException np) {
+            logger.warn("Error mandatory parameter null provided when creating student");
+            logger.warn(np.getMessage());
+            throw new ResponseStatusException(
+                    HttpStatus.BAD_REQUEST, "Error mandatory parameter null provided when creating student", np);
+        }
+        catch (TransactionSystemException cve) {
+            logger.warn(cve.getMessage());
+            throw new ResponseStatusException(
+                    HttpStatus.BAD_REQUEST, "Error Bean Validation failed", cve);
+        }
 
     }
 
 
-    @DeleteMapping("/api/cetpe/lstudent/{id}")
+    @DeleteMapping("/api/cetpe/lstudent/{username}")
     public boolean delete(@PathVariable String username) {
+
+        if (username==null || username.isEmpty() || username.length()<3 || username.length()>256) {
+            throw new ResponseStatusException(
+                    HttpStatus.BAD_REQUEST, "Error student username is not valid");
+        }
+
 
         return _learningStudentService.deleteLearningStudent(username);
